+++ src/main/java/blackjack/Application.java
@@ -5,11 +5,11 @@ import blackjack.domain.Outcome;
 import blackjack.domain.Referee;
 import blackjack.domain.card.Deck;
 import blackjack.domain.card.DeckShuffleFactory;
-import blackjack.domain.dto.DealerResponseDto;
-import blackjack.domain.dto.OutcomeDto;
-import blackjack.domain.dto.OutcomesDto;
-import blackjack.domain.dto.PlayerResponseDto;
-import blackjack.domain.dto.PlayersResponseDto;
+import blackjack.dto.DealerResponseDto;
+import blackjack.dto.OutcomeDto;
+import blackjack.dto.OutcomesDto;
+import blackjack.dto.PlayerResponseDto;
+import blackjack.dto.PlayersResponseDto;
 import blackjack.domain.gamer.Dealer;
 import blackjack.domain.gamer.Gamers;
 import blackjack.domain.gamer.Player;
