+++ src/main/java/blackjack/dto/PlayerResponseDto.java
@@ -1,4 +1,4 @@
-package blackjack.domain.dto;
+package blackjack.dto;
 
 import blackjack.domain.gamer.Name;
 import blackjack.domain.card.Card;
