+++ src/main/java/blackjack/view/mapper/SuitMapper.java
@@ -1,9 +1,9 @@
-package blackjack.view;
+package blackjack.view.mapper;
 
 import blackjack.domain.card.Suit;
 import java.util.Arrays;
 
-enum SuitTranslator {
+public enum SuitMapper {
     HEART(Suit.HEART, "하트"),
     SPADE(Suit.SPADE, "스페이드"),
     CLOVER(Suit.CLOVER, "클로버"),
@@ -12,13 +12,13 @@ enum SuitTranslator {
     private final Suit suit;
     private final String viewName;
 
-    SuitTranslator(final Suit suit, final String viewName) {
+    SuitMapper(final Suit suit, final String viewName) {
         this.suit = suit;
         this.viewName = viewName;
     }
 
-    public static String translate(final Suit suit) {
-        return Arrays.stream(SuitTranslator.values())
+    public static String mapToViewName(final Suit suit) {
+        return Arrays.stream(SuitMapper.values())
                 .filter(suitTranslator -> suitTranslator.suit == suit)
                 .findAny()
                 .orElseThrow(() -> new IllegalStateException("매칭되는 슈트가 없습니다."))
