+++ src/test/java/blackjack/view/mapper/SuitMapperTest.java
@@ -1,4 +1,4 @@
-package blackjack.view;
+package blackjack.view.mapper;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,14 +8,14 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class SuitTranslatorTest {
+class SuitMapperTest {
 
     @DisplayName("SuitTranslator는 Suit의 멤버를 모두 알고 있다.")
     @Test
     void knowAllValuesOfSuit() {
         final List<Suit> expected = List.of(Suit.values());
-        final List<Suit> actual = Arrays.stream(SuitTranslator.values())
-                .map(SuitTranslator::getSuit)
+        final List<Suit> actual = Arrays.stream(SuitMapper.values())
+                .map(SuitMapper::getSuit)
                 .toList();
 
         assertThat(actual).containsExactlyInAnyOrderElementsOf(expected);
@@ -26,7 +26,7 @@ class SuitTranslatorTest {
     void findViewName() {
         final String expected = "하트";
 
-        final String actual = SuitTranslator.translate(Suit.HEART);
+        final String actual = SuitMapper.mapToViewName(Suit.HEART);
 
         assertThat(actual).isEqualTo(expected);
     }
